requirements.txt
asgiref==3.5.2
astroid==2.11.5
colorama==0.4.4
dill==0.3.5.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
wrapt==1.14.1
PyJWT==2.4.0
# B1. virtualenv env 
# B2. .\env\scripts\activate
# B3. py -m pip install ?
Django==4.0.4
django-cors-headers==3.12.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
mysqlclient==2.1.0
# B4: py manage.py runserver